Move all exception classes to their own sub namespace/folder
